requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.6
django-admin-sortable2==2.1.10
docopt==0.6.2
docType==0.23
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2024.3.1
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
networkx==3.2.1
nltk==3.7
numpy==1.22.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipfreeze==2.0.3
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2022.6.2
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.5.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.2.1
tornado==6.4
tqdm==4.64.0
traitlets==5.14.2
transformers==4.38.2
typing_extensions==4.8.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
yarg==0.1.9
zipp==3.18.1
